requirements.txt
streamlit
requests
faiss-cpu
sentence-transformers
pypdf
